Final succsess
After few tries and fails i succeeded with uploading my code for the 10thtime! For the first time I uploaded the basic page with the home page, today I finished up on the home page with few colour changes and some design changes and added two pictures, also I started on the about page and copied the home page's content to the project and signup file just to have some data available on the buttons.